# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Langsungmakan API",
        "contact": {
            "name": "Our",
            "url": "https://www.langsungmakan.com"
        },
        "version": "1.0"
    },
    "host": "wshgay.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/customer": {
            "get": {
                "tags": [
                    "customer"
                ],
                "summary": "Retrieves the collection of Customer resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "customer collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/customer"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/customer/{id}": {
            "put": {
                "tags": [
                    "customer"
                ],
                "summary": "Retrieves the collection of Customer resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "customer collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/customer"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        }
    },
    "definitions": {
...
        "customer": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "user_id": {
                    "type": "integer"
                },
                "birthday": {
                    "type": "string",
                    "format": "date"
                },
                "gender": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "credit": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}